fix(source,ui): address PR#24 Copilot review
Four findings — one doc nit, one cross-platform compile break, two
concurrency hardenings:

1. file_special_test.go split: this file used syscall.Mkfifo /
   syscall.SockaddrUnix / etc. but had no build tag, breaking the
   Windows build at compile time (runtime GOOS skips don't help when
   the identifiers don't exist on the target). The portable mode-bit
   table test stays in file_special_test.go (no syscall import); the
   three real-syscall integration tests move to a new
   file_special_unix_test.go behind `//go:build unix`.

2. FileSource detection cache (kind/lexerName/detectErr/detected) is
   now guarded by detectMu sync.Mutex. Production shares a *FileSource
   between the Bubble Tea event loop (which calls Redetect on reload)
   and the loader's reader goroutine (which calls Open / Kind /
   Metadata via detectOnce). Pre-fix a rapid second reload could fire
   Redetect while the prior reload's still-running goroutine was
   mid-detectOnce — torn writes on the cache fields. Refactored
   detectOnce → detectOnceLocked (caller holds the lock); Kind /
   Metadata / Open / Reopen / Redetect all acquire the mutex.

3. file_other.go (Windows fallback) doc comment softened: the prior
   "already requires elevated privileges" wording implied a guarantee,
   but Developer Mode / policy can enable non-admin symlink creation.
   Now phrases as "MAY require ... depending on system configuration"
   per Copilot suggestion. Defense-in-depth claim is unchanged but no
   longer reads as a hard guarantee.
